Fixed Code Review
- 조건문에 부정문없애기
- 스프레드 문법으로 변경
- onRemove => handleRemoveTodo 변경
- Guard Clauses 사용
- Container 삭제 로직 App에서 관리
