+++ src/components/TodoTemplate.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function TodoTemplate({ children }) {
+  return (
+    <>
+      <div>To-do</div>
+      {children}
+    </>
+  );
+}
+
+export default TodoTemplate;
